requirements.txt
azure-identity
azure-mgmt-datafactory
sqlglot
python-dateutil
python-decouple
azure-synapse-artifacts